+++ libs/community/tests/unit_tests/vectorstores/test_sqlserver.py
@@ -6,7 +6,7 @@ import pytest
 from langchain_core.documents import Document
 
 from langchain_community.embeddings import FakeEmbeddings
-from langchain_community.vectorstores.azure_sql_server import SQLServer_VectorStore
+from langchain_community.vectorstores.sqlserver import SQLServer_VectorStore
 
 _CONNECTION_STRING = os.environ.get("TEST_AZURESQLSERVER_CONNECTION_STRING")
 
@@ -22,7 +22,7 @@ def store():
     yield store  # provide this data to the test
 
 
-def test_sql_server_add_texts(store) -> None:
+def test_sqlserver_add_texts(store) -> None:
     """Test that add text returns equivalent number of ids of input texts."""
     texts = ["rabbit", "cherry", "hamster", "cat", "elderberry"]
     metadatas = [
@@ -36,7 +36,7 @@ def test_sql_server_add_texts(store) -> None:
     assert len(result) == len(texts)
 
 
-def test_sql_server_add_texts_when_no_metadata_is_provided(store) -> None:
+def test_sqlserver_add_texts_when_no_metadata_is_provided(store) -> None:
     """Test that when user calls the add_texts function without providing metadata, the embedded text still get added to the vectore store."""
     texts = [
         "Good review",
@@ -48,7 +48,7 @@ def test_sql_server_add_texts_when_no_metadata_is_provided(store) -> None:
     assert len(result) == len(texts)
 
 
-def test_sql_server_add_texts_when_text_length_and_metadata_length_vary(store) -> None:
+def test_sqlserver_add_texts_when_text_length_and_metadata_length_vary(store) -> None:
     """Test that all texts provided are added into the vector store even when metadata is not available for all the texts."""
     # The text 'elderberry' and its embedded value should be added to the vector store.
     texts = ["rabbit", "cherry", "hamster", "cat", "elderberry"]
@@ -62,7 +62,7 @@ def test_sql_server_add_texts_when_text_length_and_metadata_length_vary(store) -
     assert len(result) == len(texts)
 
 
-def test_add_document_with_sql_server(store) -> None:
+def test_add_document_with_sqlserver(store) -> None:
     """Test that when add_document function is used, it integerates well with the add_text function in SQLServer Vector Store."""
     docs = [
         Document(
